move to pandautils
