--- apps/student/src/main/res/layout/bindable_assignment_spinner_item.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (C) 2022 - present Instructure, Inc.
-  ~
-  ~     This program is free software: you can redistribute it and/or modify
-  ~     it under the terms of the GNU General Public License as published by
-  ~     the Free Software Foundation, version 3 of the License.
-  ~
-  ~     This program is distributed in the hope that it will be useful,
-  ~     but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  ~     GNU General Public License for more details.
-  ~
-  ~     You should have received a copy of the GNU General Public License
-  ~     along with this program.  If not, see <http://www.gnu.org/licenses/>.
-  ~
-  -->
-<layout>
-
-    <data>
-
-        <variable
-            name="itemViewModel"
-            type="com.instructure.student.features.shareextension.target.itemviewmodels.ShareExtensionAssignmentItemViewModel" />
-    </data>
-
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center_vertical"
-        android:minHeight="45dp"
-        android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/courseName"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:ellipsize="end"
-            android:paddingLeft="8dp"
-            android:paddingRight="8dp"
-            android:singleLine="true"
-            android:text="@{itemViewModel.data.title}"
-            android:textColor="@color/textDark"
-            android:textSize="16sp" />
-    </LinearLayout>
-</layout>
