+++ libs/pandares/src/main/res/anim/ease_in_shrink.xml
