+++ libs/pandares/src/main/res/anim/expand_from_middle.xml
