+++ libs/pandares/src/main/res/anim/fab_rotate_backward.xml
