+++ libs/pandares/src/main/res/anim/fade_in_quick.xml
