+++ libs/pandares/src/main/res/anim/fade_out.xml
