+++ libs/pandares/src/main/res/anim/fade_out_quick.xml
