+++ libs/pandares/src/main/res/anim/hs_slide_out_left.xml
