+++ libs/pandares/src/main/res/anim/rotate.xml
