+++ libs/pandares/src/main/res/anim/scale_slide_in_bottom.xml
