+++ libs/pandares/src/main/res/anim/scale_slide_in_bottom_slow.xml
