+++ libs/pandares/src/main/res/anim/shrink_to_middle.xml
