+++ libs/pandares/src/main/res/anim/slide_in_from_bottom.xml
