+++ libs/pandares/src/main/res/anim/slide_in_right.xml
