+++ libs/pandares/src/main/res/anim/slide_out_to_bottom.xml
