+++ libs/pandares/src/main/res/anim/slow_push_left_in.xml
