+++ libs/pandares/src/main/res/anim/slow_push_left_out.xml
