+++ libs/pandares/src/main/res/anim/slow_push_right_in.xml
