+++ libs/pandares/src/main/res/anim/up_from_bottom.xml
