+++ libs/pandares/src/main/res/drawable/upload_file_bg.xml
